refactor: GameInning의 필드 적절한 타입으로 변경

+++ BE/baseball-be/src/main/java/kr/codesquad/baseball/inning/domain/InningRepository.java
@@ -13,11 +13,11 @@ public interface InningRepository extends CrudRepository<GameInning, Long> {
     @Query("INSERT INTO " +
             "BASEBALL.GAME_INNING(INNING, GAME_ID, TEAM_ID, SCORE) " +
             "VALUES(:inning, :gameId, :teamId, :score) ")
-    int save(@Param("inning") long inning, @Param("gameId") long gameId, @Param("teamId") long teamId, @Param("score") long score);
+    int save(@Param("inning") int inning, @Param("gameId") long gameId, @Param("teamId") long teamId, @Param("score") int score);
 
     @Modifying
     @Query("UPDATE BASEBALL.GAME_INNING " +
             "SET SCORE = :score " +
             "WHERE INNING = :inning AND GAME_ID = :gameId AND TEAM_ID = :teamId")
-    int updateScore(@Param("inning") long inning, @Param("gameId") long gameId, @Param("teamId") long teamId, @Param("score") long score);
+    int updateScore(@Param("inning") int inning, @Param("gameId") long gameId, @Param("teamId") long teamId, @Param("score") int score);
 }
